add AbstractLogging#span
